refactor(matrices): move __getitem__ to RepMatrix
